Added DSL infrastructure API detached from implementation.
Now every node aspect is described in it's own interface and
is implemented in it's own inheritor. So, if there is need for
some node logic differing from default there is easy way to
implement it.

Although, there is "gluing" class that implement node tree building
logic since API detaching is made also for encapsulation, so
internals are not accessible for DSL users, and node can't determine
where they are at node structure by themselves.